+++ src/CanDoItAll.Components.BaseLib/Components/Forms/DropDownOption.cs
@@ -1,5 +1,0 @@
-namespace CanDoItAll.Components.BaseLib;
-
-internal sealed record DropDownOption<TValue>(string Key, string Text, TValue? Value);
-
-