+++ src/CanDoItAll.Components.BaseLib/Components/Forms/FormPrimitives.cs
@@ -7,13 +7,6 @@ public enum InputLook
     Plain
 }
 
-public enum TextAreaSize
-{
-    Compact,
-    Standard,
-    Extended
-}
-
 public enum ModalSize
 {
     Compact,